This is my first idea for how to layout the payload section.
This is configured for a 6" PML outer tube and a 54mm PML motor tube in the
centre.  This payload section will be at the from of the motor tube, at the
top of the motor section.  There'll be a bulkplate above and below this
section which will be clamped together with 6mm threaded rods, thus putting
the 3d printed parts in permanent compression.